move healthcheck to shardmanager
